common/pkg/fcrdatabase/fcrdatabase_impl_v1_test.go
/*
Package fcrdatabase - handles the database inside the system.
*/
package fcrdatabase

/*
 * Copyright 2020 ConsenSys Software Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
 * the License. You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
 * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 * specific language governing permissions and limitations under the License.
 *
 * SPDX-License-Identifier: Apache-2.0
 */

import (
	"testing"

	"github.com/stretchr/testify/assert"
)

func TestPlaceholder(t *testing.T) {
	db := NewFCRDatabaseImplV1()
	db.Shutdown()
	err := db.Start()
	assert.Empty(t, err)
	defer db.Shutdown()
	err = db.Start()
	assert.NotEmpty(t, err)
}
